Relocate `/profile anilist` to `/anilist profile`
